refactor: reorganize include paths for better modularity
- Updated include paths in multiple source files to reflect new directory structure.
- Moved configuration and hardware includes to their respective directories for improved organization.
- Ensured consistency across files by aligning with the updated file structure.
